+++ v1.0.0/03_Sources/mywebsite-core/src/main/java/com/mycom/products/mywebsite/core/dao/base/InsertableDao.java
@@ -12,7 +12,7 @@ import com.mycom.products.mywebsite.core.exception.DAOException;
 import com.mycom.products.mywebsite.core.exception.DuplicatedEntryException;
 
 public interface InsertableDao<T extends BaseBean> {
-	public int insert(T record, int recordRegId) throws DuplicatedEntryException, DAOException;
+	public long insert(T record, long recordRegId) throws DuplicatedEntryException, DAOException;
 
-	public void insert(List<T> records, int recordRegId) throws DuplicatedEntryException, DAOException;
+	public void insert(List<T> records, long recordRegId) throws DuplicatedEntryException, DAOException;
 }
