+++ v1.0.0/03_Sources/mywebsite-core/src/main/java/com/mycom/products/mywebsite/core/mapper/base/UpdateableMapper.java
@@ -8,5 +8,5 @@ package com.mycom.products.mywebsite.core.mapper.base;
 import com.mycom.products.mywebsite.core.bean.BaseBean;
 
 public interface UpdateableMapper<T extends BaseBean> {
-	public int update(T record);
+	public long update(T record);
 }
